updated securities for file upload and added api and app views.py files to improve code readability
