+++ model_params.py
@@ -19,7 +19,7 @@ class params:
         folder_path = '/app/static/data/'
 
         # local
-        #folder_path = "/Users/LiamRoberts/Desktop/GalaxyWebApp/static/data/"
+        # folder_path = "/Users/LiamRoberts/Desktop/GalaxyWebApp/static/data/"
         
         # file paths
         self.folder_path = folder_path
